+++ liquid/environment.py
@@ -21,7 +21,6 @@ from liquid import builtin
 from liquid import loaders
 from liquid.analyze_tags import InnerTagMap
 from liquid.analyze_tags import TagAnalysis
-from liquid.builtin.loaders import CachingFileSystemLoader
 from liquid.context import Undefined
 from liquid.exceptions import Error
 from liquid.exceptions import LiquidSyntaxError
@@ -219,11 +218,7 @@ class Environment:
         self.mode = tolerance
 
         # Template cache
-        if (
-            cache_size
-            and cache_size > 0
-            and not isinstance(self.loader, CachingFileSystemLoader)
-        ):
+        if cache_size and cache_size > 0 and not self.loader.caching_loader:
             self.cache: Optional[MutableMapping[Any, Any]] = LRUCache(cache_size)
             self.auto_reload = auto_reload
         else: