+++ CHANGES.md
@@ -11,7 +11,7 @@
 
 - Added context-aware template loader bootstrapping methods. See [#109](https://github.com/jg-rp/liquid/pull/109).
 - Added support for async class-based filters. See [#115](https://github.com/jg-rp/liquid/pull/115).
-- Added `CachingFileSystemLoader`, a template loader that handles its own cache rather than relying on the `Environment` template cache, which can't handle context-aware loaders.
+- Added `CachingFileSystemLoader`, a template loader that handles its own cache rather than relying on the `Environment` template cache, which can't handle context-aware loaders. See [#116](https://github.com/jg-rp/liquid/pull/116).
 
 ## Version 1.8.1